Move safe URL building utilities into core, so we can reuse them in the client
